pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drug-discovery-openenv"
version = "1.0.0"
description = "OpenEnv lead optimization: SMILES, RDKit descriptors, structured medicinal-chemistry actions, multi-task grading."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "fastapi>=0.115.0",
  "openai>=1.66.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.0",
  "uvicorn>=0.34.0",
  "rdkit>=2023.3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
server = "server.app:main"
drug-discovery-server = "server.app:main"
drug-discovery-baseline = "drug_discovery_env.baseline:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["drug_discovery_env*", "server*"]

[tool.setuptools.package-data]
drug_discovery_env = ["tasks/*.json"]
